Simplify the doc annotations.
This removes all but the lint internal features, making it so `doc_autocfg` can work as expected, and avoiding disconnects between the docs and the required features. This cannot have backward-incompatible issues since we forced compile errors if incorrect feature combinations were used.
