+++ extras/benchmark/parse-float/Cargo.toml
@@ -30,8 +30,8 @@ power-of-two = ["lexical-util/power-of-two", "lexical-parse-float/power-of-two"]
 format = ["lexical-util/format", "lexical-parse-float/format"]
 compact = ["lexical-util/compact", "lexical-parse-float/compact"]
 asm = []
-integers = ["lexical-util/integers"]
-floats = ["lexical-util/floats"]
+floats = ["lexical-util/parse-floats"]
+integers = []  # needed for feature detection
 json = []
 
 [[bench]]
