+++ extras/benchmark/write-integer/Cargo.toml
@@ -30,8 +30,8 @@ radix = ["lexical-util/radix", "lexical-write-integer/radix"]
 power-of-two = ["lexical-util/power-of-two", "lexical-write-integer/power-of-two"]
 format = ["lexical-util/format", "lexical-write-integer/format"]
 compact = ["lexical-util/compact", "lexical-write-integer/compact"]
-integers = ["lexical-util/integers"]
-floats = ["lexical-util/floats"]
+floats = []  # needed for feature detection
+integers = ["lexical-util/write-integers"]
 json = []
 
 [[bench]]
