+++ extras/util/Cargo.toml
@@ -27,14 +27,10 @@ std = ["lexical-util/std"]
 power-of-two = ["lexical-util/power-of-two"]
 radix = ["lexical-util/radix"]
 format = ["lexical-util/format"]
-write-integers = ["lexical-util/write-integers", "write", "integers"]
-write-floats = ["lexical-util/write-floats", "write", "floats"]
-parse-integers = ["lexical-util/parse-integers", "parse", "integers"]
-parse-floats = ["lexical-util/parse-floats", "parse", "floats"]
+write-integers = ["lexical-util/write-integers"]
+write-floats = ["lexical-util/write-floats"]
+parse-integers = ["lexical-util/parse-integers"]
+parse-floats = ["lexical-util/parse-floats"]
 compact = ["lexical-util/compact"]
 f16 = ["lexical-util/f16", "parse-floats", "write-floats"]
 lint = ["lexical-util/lint"]
-write = ["lexical-util/write"]
-parse = ["lexical-util/parse"]
-integers = ["lexical-util/integers"]
-floats = ["lexical-util/floats"]
