+++ lexical-core/tests/issue_97_tests.rs
@@ -1,4 +1,4 @@
-#![cfg(all(feature = "parse", feature = "format"))]
+#![cfg(all(any(feature = "parse-floats", feature = "parse-integers"), feature = "format"))]
 
 use core::num;
