+++ lexical-util/src/num.rs
@@ -4,14 +4,14 @@
 //! types, and trait bounds, and conversions for working with
 //! numbers in generic code.
 
-#![cfg_attr(any(), rustfmt::skip)]
-
 use core::{fmt, mem, ops};
 
 #[cfg(feature = "f16")]
 use crate::bf16::bf16;
 #[cfg(feature = "f16")]
 use crate::f16::f16;
+#[cfg(all(not(feature = "std"), any(feature = "parse-floats", feature = "write-floats")))]
+use crate::libm;
 
 // AS PRIMITIVE
 // ------------
@@ -852,8 +852,7 @@ unsigned_integer_impl! { u8 u16 u32 u64 u128 usize }
 /// exponent of base 2.
 ///
 /// [`floats`]: https://en.wikipedia.org/wiki/Floating-point_arithmetic
-#[cfg(feature = "floats")]
-#[cfg_attr(docsrs, doc(cfg(any(feature = "parse-floats", feature = "write-floats"))))]
+#[cfg(any(feature = "parse-floats", feature = "write-floats"))]
 pub trait Float: Number + ops::Neg<Output = Self> {
     /// Unsigned type of the same size.
     type Unsigned: UnsignedInteger;
@@ -998,7 +997,8 @@ pub trait Float: Number + ops::Neg<Output = Self> {
         self.to_bits() & Self::EXPONENT_MASK == Self::Unsigned::ZERO
     }
 
-    /// Returns true if the float is NaN, positive infinity, or negative infinity.
+    /// Returns true if the float is NaN, positive infinity, or negative
+    /// infinity.
     #[inline(always)]
     fn is_special(self) -> bool {
         self.to_bits() & Self::EXPONENT_MASK == Self::EXPONENT_MASK
@@ -1180,7 +1180,7 @@ pub trait Float: Number + ops::Neg<Output = Self> {
 }
 
 /// Define the float literals.
-#[cfg(feature = "floats")]
+#[cfg(any(feature = "parse-floats", feature = "write-floats"))]
 macro_rules! float_literals {
     ($float:ty) => {
         const ZERO: $float = 0.0;
@@ -1196,7 +1196,7 @@ macro_rules! float_literals {
 }
 
 /// Define the float masks.
-#[cfg(feature = "floats")]
+#[cfg(any(feature = "parse-floats", feature = "write-floats"))]
 macro_rules! float_masks {
     (
         float =>
@@ -1374,7 +1374,7 @@ impl Float for bf16 {
     }
 }
 
-#[cfg(feature = "floats")]
+#[cfg(any(feature = "parse-floats", feature = "write-floats"))]
 impl Float for f32 {
     type Unsigned = u32;
     float_literals!(f32);
@@ -1407,7 +1407,7 @@ impl Float for f32 {
         return f32::ln(self);
 
         #[cfg(not(feature = "std"))]
-        return logf(self);
+        return libm::logf(self);
     }
 
     #[inline(always)]
@@ -1416,7 +1416,7 @@ impl Float for f32 {
         return f32::floor(self);
 
         #[cfg(not(feature = "std"))]
-        return floorf(self);
+        return libm::floorf(self);
     }
 
     #[inline(always)]
@@ -1430,7 +1430,7 @@ impl Float for f32 {
     }
 }
 
-#[cfg(feature = "floats")]
+#[cfg(any(feature = "parse-floats", feature = "write-floats"))]
 impl Float for f64 {
     type Unsigned = u64;
     float_literals!(f64);
@@ -1463,7 +1463,7 @@ impl Float for f64 {
         return f64::ln(self);
 
         #[cfg(not(feature = "std"))]
-        return logd(self);
+        return libm::logd(self);
     }
 
     #[inline(always)]
@@ -1472,7 +1472,7 @@ impl Float for f64 {
         return f64::floor(self);
 
         #[cfg(not(feature = "std"))]
-        return floord(self);
+        return libm::floord(self);
     }
 
     #[inline(always)]
@@ -1503,276 +1503,3 @@ impl Float for f64 {
 //     const DENORMAL_EXPONENT: i32 = 1 - Self::EXPONENT_BIAS;
 //     const MAX_EXPONENT: i32 = 0x7FFF - Self::EXPONENT_BIAS;
 // }
-
-// FLOAT HELPERS
-// -------------
-
-// These are adapted from libm, a port of musl libc's libm to Rust.
-// libm can be found online [here](https://github.com/rust-lang/libm),
-// and is similarly licensed under an Apache2.0/MIT license
-
-/// # Safety
-///
-/// Safe as long as `e` is properly initialized.
-#[cfg(all(not(feature = "std"), feature = "floats"))]
-macro_rules! volatile {
-($e:expr) => {
-    // SAFETY: safe as long as `$e` has been properly initialized.
-    unsafe {
-        core::ptr::read_volatile(&$e);
-    }
-};
-}
-
-/// Floor (f64)
-///
-/// Finds the nearest integer less than or equal to `x`.
-#[cfg(all(not(feature = "std"), feature = "floats"))]
-fn floord(x: f64) -> f64 {
-    const TOINT: f64 = 1. / f64::EPSILON;
-
-    let ui = x.to_bits();
-    let e = ((ui >> 52) & 0x7ff) as i32;
-
-    if (e >= 0x3ff + 52) || (x == 0.) {
-        return x;
-    }
-    /* y = int(x) - x, where int(x) is an integer neighbor of x */
-    let y = if (ui >> 63) != 0 {
-        x - TOINT + TOINT - x
-    } else {
-        x + TOINT - TOINT - x
-    };
-    /* special case because of non-nearest rounding modes */
-    if e < 0x3ff {
-        volatile!(y);
-        return if (ui >> 63) != 0 {
-            -1.
-        } else {
-            0.
-        };
-    }
-    if y > 0. {
-        x + y - 1.
-    } else {
-        x + y
-    }
-}
-
-/// Floor (f32)
-///
-/// Finds the nearest integer less than or equal to `x`.
-#[cfg(all(not(feature = "std"), feature = "floats"))]
-fn floorf(x: f32) -> f32 {
-    let mut ui = x.to_bits();
-    let e = (((ui >> 23) as i32) & 0xff) - 0x7f;
-
-    if e >= 23 {
-        return x;
-    }
-    if e >= 0 {
-        let m: u32 = 0x007fffff >> e;
-        if (ui & m) == 0 {
-            return x;
-        }
-        volatile!(x + f32::from_bits(0x7b800000));
-        if ui >> 31 != 0 {
-            ui += m;
-        }
-        ui &= !m;
-    } else {
-        volatile!(x + f32::from_bits(0x7b800000));
-        if ui >> 31 == 0 {
-            ui = 0;
-        } else if ui << 1 != 0 {
-            return -1.0;
-        }
-    }
-    f32::from_bits(ui)
-}
-
-/* origin: FreeBSD /usr/src/lib/msun/src/e_log.c */
-/*
- * ====================================================
- * Copyright (C) 1993 by Sun Microsystems, Inc. All rights reserved.
- *
- * Developed at SunSoft, a Sun Microsystems, Inc. business.
- * Permission to use, copy, modify, and distribute this
- * software is freely granted, provided that this notice
- * is preserved.
- * ====================================================
- */
-/* log(x)
- * Return the logarithm of x
- *
- * Method :
- *   1. Argument Reduction: find k and f such that
- *                      x = 2^k * (1+f),
- *         where  sqrt(2)/2 < 1+f < sqrt(2) .
- *
- *   2. Approximation of log(1+f).
- *      Let s = f/(2+f) ; based on log(1+f) = log(1+s) - log(1-s)
- *               = 2s + 2/3 s**3 + 2/5 s**5 + .....,
- *               = 2s + s*R
- *      We use a special Remez algorithm on [0,0.1716] to generate
- *      a polynomial of degree 14 to approximate R The maximum error
- *      of this polynomial approximation is bounded by 2**-58.45. In
- *      other words,
- *                      2      4      6      8      10      12      14
- *          R(z) ~ Lg1*s +Lg2*s +Lg3*s +Lg4*s +Lg5*s  +Lg6*s  +Lg7*s
- *      (the values of Lg1 to Lg7 are listed in the program)
- *      and
- *          |      2          14          |     -58.45
- *          | Lg1*s +...+Lg7*s    -  R(z) | <= 2
- *          |                             |
- *      Note that 2s = f - s*f = f - hfsq + s*hfsq, where hfsq = f*f/2.
- *      In order to guarantee error in log below 1ulp, we compute log
- *      by
- *              log(1+f) = f - s*(f - R)        (if f is not too large)
- *              log(1+f) = f - (hfsq - s*(hfsq+R)).     (better accuracy)
- *
- *      3. Finally,  log(x) = k*ln2 + log(1+f).
- *                          = k*ln2_hi+(f-(hfsq-(s*(hfsq+R)+k*ln2_lo)))
- *         Here ln2 is split into two floating point number:
- *                      ln2_hi + ln2_lo,
- *         where n*ln2_hi is always exact for |n| < 2000.
- *
- * Special cases:
- *      log(x) is NaN with signal if x < 0 (including -INF) ;
- *      log(+INF) is +INF; log(0) is -INF with signal;
- *      log(NaN) is that NaN with no signal.
- *
- * Accuracy:
- *      according to an error analysis, the error is always less than
- *      1 ulp (unit in the last place).
- *
- * Constants:
- * The hexadecimal values are the intended ones for the following
- * constants. The decimal values may be used, provided that the
- * compiler will convert from decimal to binary accurately enough
- * to produce the hexadecimal values shown.
- */
-
-#[allow(clippy::eq_op, clippy::excessive_precision)] // reason="values need to be exact under all conditions"
-#[cfg(all(not(feature = "std"), feature = "floats"))]
-fn logd(mut x: f64) -> f64 {
-    const LN2_HI: f64 = 6.93147180369123816490e-01; /* 3fe62e42 fee00000 */
-    const LN2_LO: f64 = 1.90821492927058770002e-10; /* 3dea39ef 35793c76 */
-    const LG1: f64 = 6.666666666666735130e-01; /* 3FE55555 55555593 */
-    const LG2: f64 = 3.999999999940941908e-01; /* 3FD99999 9997FA04 */
-    const LG3: f64 = 2.857142874366239149e-01; /* 3FD24924 94229359 */
-    const LG4: f64 = 2.222219843214978396e-01; /* 3FCC71C5 1D8E78AF */
-    const LG5: f64 = 1.818357216161805012e-01; /* 3FC74664 96CB03DE */
-    const LG6: f64 = 1.531383769920937332e-01; /* 3FC39A09 D078C69F */
-    const LG7: f64 = 1.479819860511658591e-01; /* 3FC2F112 DF3E5244 */
-
-    let x1p54 = f64::from_bits(0x4350000000000000); // 0x1p54 === 2 ^ 54
-
-    let mut ui = x.to_bits();
-    let mut hx: u32 = (ui >> 32) as u32;
-    let mut k: i32 = 0;
-
-    if (hx < 0x00100000) || ((hx >> 31) != 0) {
-        /* x < 2**-126 */
-        if ui << 1 == 0 {
-            return -1. / (x * x); /* log(+-0)=-inf */
-        }
-        if hx >> 31 != 0 {
-            return (x - x) / 0.0; /* log(-#) = NaN */
-        }
-        /* subnormal number, scale x up */
-        k -= 54;
-        x *= x1p54;
-        ui = x.to_bits();
-        hx = (ui >> 32) as u32;
-    } else if hx >= 0x7ff00000 {
-        return x;
-    } else if hx == 0x3ff00000 && ui << 32 == 0 {
-        return 0.;
-    }
-
-    /* reduce x into [sqrt(2)/2, sqrt(2)] */
-    hx += 0x3ff00000 - 0x3fe6a09e;
-    k += ((hx >> 20) as i32) - 0x3ff;
-    hx = (hx & 0x000fffff) + 0x3fe6a09e;
-    ui = ((hx as u64) << 32) | (ui & 0xffffffff);
-    x = f64::from_bits(ui);
-
-    let f: f64 = x - 1.0;
-    let hfsq: f64 = 0.5 * f * f;
-    let s: f64 = f / (2.0 + f);
-    let z: f64 = s * s;
-    let w: f64 = z * z;
-    let t1: f64 = w * (LG2 + w * (LG4 + w * LG6));
-    let t2: f64 = z * (LG1 + w * (LG3 + w * (LG5 + w * LG7)));
-    let r: f64 = t2 + t1;
-    let dk: f64 = k as f64;
-    s * (hfsq + r) + dk * LN2_LO - hfsq + f + dk * LN2_HI
-}
-
-/* origin: FreeBSD /usr/src/lib/msun/src/e_logf.c */
-/*
- * Conversion to float by Ian Lance Taylor, Cygnus Support, ian@cygnus.com.
- */
-/*
- * ====================================================
- * Copyright (C) 1993 by Sun Microsystems, Inc. All rights reserved.
- *
- * Developed at SunPro, a Sun Microsystems, Inc. business.
- * Permission to use, copy, modify, and distribute this
- * software is freely granted, provided that this notice
- * is preserved.
- * ====================================================
- */
-
-#[allow(clippy::eq_op, clippy::excessive_precision)] // reason="values need to be exact under all conditions"
-#[cfg(all(not(feature = "std"), feature = "floats"))]
-fn logf(mut x: f32) -> f32 {
-    const LN2_HI: f32 = 6.9313812256e-01; /* 0x3f317180 */
-    const LN2_LO: f32 = 9.0580006145e-06; /* 0x3717f7d1 */
-    /* |(log(1+s)-log(1-s))/s - Lg(s)| < 2**-34.24 (~[-4.95e-11, 4.97e-11]). */
-    const LG1: f32 = 0.66666662693; /* 0xaaaaaa.0p-24 */
-    const LG2: f32 = 0.40000972152; /* 0xccce13.0p-25 */
-    const LG3: f32 = 0.28498786688; /* 0x91e9ee.0p-25 */
-    const LG4: f32 = 0.24279078841; /* 0xf89e26.0p-26 */
-
-    let x1p25 = f32::from_bits(0x4c000000); // 0x1p25f === 2 ^ 25
-
-    let mut ix = x.to_bits();
-    let mut k = 0i32;
-
-    if (ix < 0x00800000) || ((ix >> 31) != 0) {
-        /* x < 2**-126 */
-        if ix << 1 == 0 {
-            return -1. / (x * x); /* log(+-0)=-inf */
-        }
-        if (ix >> 31) != 0 {
-            return (x - x) / 0.; /* log(-#) = NaN */
-        }
-        /* subnormal number, scale up x */
-        k -= 25;
-        x *= x1p25;
-        ix = x.to_bits();
-    } else if ix >= 0x7f800000 {
-        return x;
-    } else if ix == 0x3f800000 {
-        return 0.;
-    }
-
-    /* reduce x into [sqrt(2)/2, sqrt(2)] */
-    ix += 0x3f800000 - 0x3f3504f3;
-    k += ((ix >> 23) as i32) - 0x7f;
-    ix = (ix & 0x007fffff) + 0x3f3504f3;
-    x = f32::from_bits(ix);
-
-    let f = x - 1.;
-    let s = f / (2. + f);
-    let z = s * s;
-    let w = z * z;
-    let t1 = w * (LG2 + w * LG4);
-    let t2 = z * (LG1 + w * LG3);
-    let r = t2 + t1;
-    let hfsq = 0.5 * f * f;
-    let dk = k as f32;
-    s * (hfsq + r) + dk * LN2_LO - hfsq + f + dk * LN2_HI
-}
